CHROMIUM: drm/print: Don't add events to drm trace instance
We only use the drm trace instance for printing drm logs to
/sys/kernel/tracing/instances/drm/trace, not for event tracing. Yet we
needlessly populate /sys/kernel/tracing/instances/drm/events with all
the same events as the global tracer (so the vast majority of those
events have nothing to do with drm).

As a result we have 2x more files in tracefs, which in particular
results in ~14MB more slab memory consumed for tracefs inodes in
tracefs_create_file().

So disable adding any events to the drm instance.

BUG=b:309475612
UPSTREAM-TASK=b:311737594
TEST=1. No event directories in /sys/kernel/tracing/instances/drm/events
2. In /proc/meminfo SReclaimable is 12-14MB lower than w/o this CL

Change-Id: I7d80aabfeae4c1cdfa3ac0a5986063108c71573c
Signed-off-by: Dmytro Maluka <dmaluka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5132218
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Dmytro Maluka <dmaluka@google.com>