BACKPORT: FROMGIT: bpf: Support uid and gid when mounting bpffs
Parse uid and gid in bpf_parse_param() so that they can be passed in as
the `data` parameter when mount() bpffs. This will be useful when we
want to control which user/group has the control to the mounted bpffs,
otherwise a separate chown() call will be needed.

Signed-off-by: Jie Jiang <jiejiang@chromium.org>
Signed-off-by: Andrii Nakryiko <andrii@kernel.org>
Acked-by: Mike Frysinger <vapier@chromium.org>
Acked-by: Christian Brauner <brauner@kernel.org>
Acked-by: Andrii Nakryiko <andrii@kernel.org>
Link: https://lore.kernel.org/bpf/20231212093923.497838-1-jiejiang@chromium.org
(cherry picked from commit 750e785796bb72423b97cac21ecd0fa3b3b65610
 https://git.kernel.org/pub/scm/linux/kernel/git/bpf/bpf-next.git for-next)

Conflicts:
   include/linux/bpf.h
   kernel/bpf/inode.c

The conflicts are mainly because in the upstream code,
`struct bpf_mount_opts` touched in this patch was moved from inode.c to
bpf.h.

BUG=b:312117303
TEST=tast run $DUT network.QosWebRTC
TEST=`findmnt -T /sys/fs/bpf` shows the correct uid and gid

Change-Id: I235d85c85cd4e92ff131c8299696255dec6444d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5058379
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>