UPSTREAM: wifi: mac80211: Skip association timeout update after comeback rejection
It is possible for the TX status report for the (Re)Association Request
frame to be delayed long enough for the AP's (Re)Association Response
frame to be received and processed before it. If that were to happen for
a case where the AP rejects the association with indication to come back
later, the association timeout and retry state should not be modified
anymore with the TX status information that would be processed after
this. Updating the association timeout in such a reverse order of events
could result in shortening the timeouts for the association comeback
mechanism and that could result in the association failing.

Track whether we have already processed association rejection with
comeback time and if so, skip the timeout and retry update on any
following TX status report.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Link: https://msgid.link/20231219174814.2581575-1-j@w1.fi
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit b133fdf07db849eb2ce25df19d88ff31701d3b6d)

BUG=b:394323596
TEST=tast.wifi.PMKSACaching.ftsae

Change-Id: I27c80d35e2f25c7424374dce41d4efced14ee383
Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6262067
Reviewed-by: Jun Yu <junyuu@chromium.org>
Tested-by: Tim Lee <timlee@realtek.com>
Commit-Queue: Ruth Mekonnen <rmekonnen@google.com>
Reviewed-by: Ruth Mekonnen <rmekonnen@google.com>