CHROMIUM: config: armel: Rename chromiumos-arm.flavour.config
chromiumos-<arch>.flavour.config and common.config can be ambiguous
sometimes if they are only referenced by their names.

Append `-generic` to the former and hopefully make it clearer that it's
for building a generic kernel, not for putting common kernel configs
that will be applied to all platforms.

BUG=b:317315065
TEST=emerge-arm-generic chromeos-kernel-$version

Cq-Depend: chromium:5151754
Change-Id: Ib026d3eb6d789b58e9a881301fe5a7497e93ac56
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5175050
Reviewed-by: Douglas Anderson <dianders@chromium.org>