UPSTREAM: arm64: dts: mt8183: kukui-jacuzzi: Drop bogus anx7625 panel_flag property
The panel_flag property was used in ChromeOS's downstream kernel. It was
used to signal whether the downstream device was a fixed panel or
a connector for an external display.

This property was dropped in favor of standard OF graph descrptions of
downstream display panels and bridges.

Drop the property from the device tree file.

Fixes: cabc71b08eb5 ("arm64: dts: mt8183: Add kukui-jacuzzi-damu board")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Link: https://lore.kernel.org/r/20231130074032.913511-2-wenst@chromium.org
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
(cherry picked from commit 561003e16490760042c6fd6e8880d9406916a9e5)

BUG=b:184019253
TEST=boot juniper 5.10

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Change-Id: I412b3a97b1d4d498526cdec8aaa57106933c35ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5199214
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>