CHROMIUM: Fix le and bredr concurrent connection prevention
When connecting to LE device, we might need to first scan the
advertisement before initiating the connection request. Therefore, at
this point we already queue the connection.

Since we have the logic to not connect multiple concurrent connections,
it is possible to confuse this connection with itself when checking the
connection queue, therefore preventing any connection at all.

Add some logic to distinguish this case.

BUG=b:322563796
TEST=Pass CTSV BLE Secure Client Test on an Arc-R device
UPSTREAM-TASK=b:315919954

Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Change-Id: Ieeaa56b274c211007345599d15cbdb1157bbbcf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5257163
Reviewed-by: Yun-hao Chung <howardchung@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>