CHROMIUM: iwl7000: mvm: fix the TXF mapping for BZ devices
Those devices' fifos are numbered differently.
Because of that, we were looking at the size of the VO fifo size to
determine the size of the A-MSDU which led to a lower throughput.

Note that for those devices the only user of the AC -> fifo mapping is
the size limitation of A-MSDU.

Change-Id: I73e44d5fc474ebb6f275b9008950e59c012f33b2
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 8487a1cc8cb29daa9ce5af8fad330aa2f64c6c69