BACKPORT: ANDROID: power: wakeup_reason: Move back to upstream logging for suspend aborts
This is a direct revert of a portion of the ANDROID patch: "REVISIT:
ANDROID: power: wakeup_reason: add an API to log wakeup
reasons". Specifically, it was generated like this on the
chromeos-6.6 tree:

  git checkout c810b19c3dc2 # REVISIT: ANDROID: power: wakeup_reas...
  git checkout HEAD~ -- kernel/power/process.c
  git commit -m "make the patch"
  git format-patch HEAD~

...with a very slight merge resolution since the ChromeOS tree
resolved some merges a little differently. Final confirmation shows no
diffs from kernel v6.6 in the file "kernel/power/process.c"

That portion of the patch appears to have been squashed in but was
originally part of a patch with this description:

  ANDROID: power: wakeup_reason: Avoids bogus error messages for the suspend aborts.

  Avoids printing bogus error message "tasks refusing to freeze", in
  cases where pending wakeup source caused the suspend abort.

While indeed you could see how someone could get confused by the
implication that the printed tasks _caused_ the abort, the lack of the
printed tasks can cause it to be difficult to debug some issues.
Specifically the set of tasks may be quite interesting if we waited a
long time before aborting.

As an example, the ChromeOS test labs usually stress test
suspend/resume using an RTC alarm that's less than 20 seconds. This
means that in every case we'll see an "aborted" suspend (and no task
info) even if the real problem was some task taking a long time to
suspend. In analyzing one such failure I saw the message "Freezing of
tasks aborted after 15.359 seconds" with no task information even
though clearly some task was causing us problems.

Conflicts:
	kernel/power/process.c
This is a direct pick of the chromeos-5.15 patch. Diffing the file
against v5.4.268 does show diffs due to another ANDROID patch but
looks expected.

BUG=b:324958258
TEST=See TEST-ONLY testing patch https://crrev.com/c/5287829/

Bug: 324958258
Change-Id: I0a5dfe8daf261ac49ce2c0704d1ce2a517848d16
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 272d889f945254e061b48423043659d2c12e7163
 https://android.googlesource.com/kernel/common android-mainline)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5297234
Reviewed-by: Guenter Roeck <groeck@chromium.org>