CHROMIUM: iwl7000: mvm: Add helper functions to update ESR status
There are reasons for which we need to exit esr, but not to block it
completely, and there are reasons for which we need to block esr.

For both reason types we have the enum iwl_mvm_esr_state, when the
blocking reasons are stored in the `mvmvif::esr_disable_reason` bitmap.

This change introduces the APIs to use in the different cases:
- iwl_mvm_exit_esr - will exit from eSR mode.
- iwl_mvm_block_esr - will update the bitmap and exit esr, to
  be used for the blocking reasons only.
- iwl_mvm_unblock_esr - will update the bitmap. To be used for the
  blocking reasons only.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I552926065521f5f848c37b0bd845494bd7865fb7
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b639914bcf5350a04982dc9e53ff7d0f055fa5bc