CHROMIUM: drm/i915: Pick the highest bandwidth mode when multiple are preferred
If a panel has multiple preferred modes, the driver will select the
first it encounters. If this happens to be the downclock mode, we will
end up driving the panel at the incorrect refresh rate. Fix this by
selecting the highest bandwidth preferred mode available.

This behavior is fixed in 5.15+ kernels by the DRRS patchsets. Those
would be too disruptive/risky to backport, so pull back this small fix
to reduce risk.

BUG=b:229061262
UPSTREAM-TASK=b:326143024
TEST=Tested on delbing SKU1 with 165Hz panel, confirmed refresh is
correct

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I384c34ac8377cc7e9815b6dc6386e78b661e15ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5311454
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Manasi Navare <navaremanasi@chromium.org>