CHROMIUM: drm/mediatek: Disable broken gamma ramp
Recent Chrome changes enable use of the gamma ramp on Mediatek, but it
turns out that it's broken. Let's disable it until it gets fixed since
this isn't a hard requirement for older devices (on top of the fact
that we don't really need it, this bug seems to be fixed upstream as
part of the 8195 support work).

BUG=b:326134132
TEST=tested on asurada, no yellow rendering
UPSTREAM-TASK=b:326146692

Change-Id: Id7ec00322f51e6a79ad278e2332bc26161763e17
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5312003
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>