UPSTREAM: drm/mediatek: Declare Z Position for all planes
MediaTek SoCs support multiple planes, one of which is the primary
and all the others are overlays (and CURSOR is the last overlay).

In all currently supported SoCs, the Z order of the overlays can't
be changed with any fast muxing action, and can only be changed by
swapping the contents of the entire register set of one overlay
with the other to internally reorder the layer properties, which
is indeed feasible, but probably more expensive than desired.

Declare the Z position for all planes with an immutable property
at least for now, so that the userspace can take its decisions
accordingly.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: Fei Shao <fshao@chromium.org>
Acked-by: Daniel Stone <daniels@collabora.com>
Reviewed-by: CK Hu <ck.hu@mediatek.com>
Link: https://patchwork.kernel.org/project/dri-devel/patch/20240718082507.216764-1-angelogioacchino.delregno@collabora.com/
Signed-off-by: Chun-Kuang Hu <chunkuang.hu@kernel.org>
(cherry-picked from commit 103b90752f3dda74abc56993390fa303147e7fbc)

BUG=b:258560987
TEST=TEST=emerge-rauru sys-kernel/chromeos-kernel-6_6
UPSTREAM-TASK=b:287363873

Change-Id: Ifb9e0fb71cb09f896dd7b73a123dd2119ccfa4b9
Signed-off-by: Jason-jh Lin <jason-jh.lin@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6109987
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>