CHROMIUM: iwl7000: mld: Implement driver Rx queues sync command
In multi rx queue HW, without execessive locking, there is no sync
between the ctrl path (default queue) and the rest of the rx queues.

This might cause issues on certain situations. For example, in case
a delBA was processed on a default queue but out of order packets
still wait for processing on the other queue.
In addition, mac80211 will call the driver whenever there is a race
between RSS queues and control path that requires a processing of all
pending frames in RSS queues.

The solution is to implement internal messaging between the control
path and the other RX queues. The driver sends a message to the
firmware, which echoes it to all relevant queues. The driver then
waits for the notification to be acknowledged on all queues. Since
the notification is inserted into the RX queue, the RX path will
process all pending frames before handling the notification.
This ensures synchronization and avoids race conditions between the
control path and RSS queues.

This patch implements the API for sending the command; the next patch
will handle sending the command and handling the notification.

Signed-off-by: Daniel Gabay <daniel.gabay@intel.com>
Change-Id: I5f6df13d61b29288e44c1c6e70c62ff023a33c21
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: b0da1d348d016341afb908a0cb711fd4ad9c08b7