CHROMIUM: iwl7000: mld: avoid a warning in recovery
When we have an assert just before adding a station, mac80211 will
continue with adding it to the driver before we even get to the reconfig
flow.
But since the in_restart bit will already be set then,
we skip the allocation of the fw sta id, leading to
the warning in iwl_mld_add_modify_sta_cmd about an invalid fw sta id.

In reconfig, a station is re-added, so we really need to skip the ID
allocation, as we already have one,
but in error-before-recovey, we are adding the station for the first
time, so there is no a pre-allocated id. Hence the warning.

Fix this by not adding the link sta at all if we are in
error-before-recovery. The command will fail anyway.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I9ad8a5460bba33e7e4e4e2267005a0c912ea601b
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: afc1a35294625c913e45a0dbec2a586dd451bf82