CHROMIUM: iwl7000: mld: don't deactivate the link on its removal
Currently, iwl_mld_remove_link will also deactivate the link if it is
marked as active.

If (due to another bug) the link is marked as active when it is actually
not, we will try to dereference the chanctx pointer in the link when it
is invalid - and we'll crash.

To better detect such case, don't activate the link in
iwl_mld_remove_link. Instead, warn if the link is active.

The user should know whether the link is supposed to be active - and
then call iwl_mld_deactivate_link first, and only then call
iwl_mld_remvoe_link.

In cases where the link is not supposed to be active - we should call
iwl_mld_remove_link and get the warning if it is, instead of getting to
the crash described above.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Change-Id: I66863ac5c2d428bc314cd33b2f468d7cc332088f
tested: iil_jenkins iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: c21a1ff65d21e82b3b5d6c2dc3a2079a2934ca2a