CHROMIUM: iwl7000: rework firmware error handling
In order to later add the ability to do deeper resets of the
device when it crashes, first restructure the firmware error
handling. Instead of having just a single nic_error() method
that handles all, split it:
 - nic_error() just handles and prints the error itself,
 - dump_error() synchronously creates an error dump, and
 - sw_reset() will be called to request doing a SW reset.

This changes the architecture so that the transport is now
responsible for deciding how to do the reset, and therefore
the handling of reprobe if error occurs during reconfig
moves there, which necessitates adding a method there that
notifies the transport that the recovery was completed.

Actually introducing the model under which deeper resets can
be done will be in future patches.

Change-Id: I96a9243e7877808ed6d1bff6967c15d6c24882f0
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: bab8ab61ff02efaf895ed83703596924d12b8974