CHROMIUM: iwl7000: mvm: restart device through NMI
When some channel context manipulations fail, the device
is going to be restarted to try to recover. Make this go
through a real FW restart via an NMI so the transport is
aware of it and can later handle escalation, and to make
it easier to restructure the code later.

Change-Id: I2e729f402db58a76cea620b6f62a02da49a10b48
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 10af1ed41399046e74e4217bfd4e6ae7518729d6